+++ docs/CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project (hashcat.launcher) will be documented in thi
  - Missing Started and ETA info for tasks
  - Task guess info (guess base and guess mod) (#53)
 
+### Changed:
+ - Disable potfile by default
+
 ## [1.0.0-beta] 20/12/2021
 ### Changed:
  - Exhausted status color to pink (#58)